Read-only t/core_rwops.t
t/core_rwops.t used to create a file in the current working directory
and then test that SDL::RWOps can open it, seek in it and read it.
I.e. the test did not perform any write operation, yet setting up the
test required writing.

This unnecessary write prevented from running the tests from
a read-only location.

This patch precreates the test file so that no write operation is
needed when running the test.